+++ site/pages/index.js
@@ -25,13 +25,9 @@ export default class Home extends React.Component {
     return (
       <Page path="/">
         <div className="pt24">
-          <div className="flex-parent-ml flex-parent--wrap-ml">
-            <h1 className="flex-child-ml flex-child--grow-ml txt-h2 txt-bold">
-              Assembly.css
-            </h1>
-            <div className="flex-child-ml mt18 color-darken50">
-              v{pkg.version}
-            </div>
+          <div className="flex-ml flex--wrap-ml">
+            <h1 className="flex-child-grow-ml txt-h2 txt-bold">Assembly.css</h1>
+            <div className="mt18 color-darken50">v{pkg.version}</div>
           </div>
           <p className="txt-l mt18">
             Assembly is an atomic CSS framework with great default form element
@@ -134,8 +130,8 @@ export default class Home extends React.Component {
             Overview
           </h2>
 
-          <div className="mt60 flex-parent-mm">
-            <div className="flex-child flex-child--no-shrink mr24 mb0-mm mb18 w60 h60">
+          <div className="mt60 flex-mm">
+            <div className="flex-child-no-shrink mr24 mb0-mm mb18 w60 h60">
               <img src={modifierImg} />
             </div>
             <div>
@@ -158,8 +154,8 @@ export default class Home extends React.Component {
             </div>
           </div>
 
-          <div className="mt60 flex-parent-mm">
-            <div className="flex-child flex-child--no-shrink mr24 mb0-mm mb18 w60 h60">
+          <div className="mt60 flex-mm">
+            <div className="flex-child-no-shrink mr24 mb0-mm mb18 w60 h60">
               <img src={baselineGridImg} />
             </div>
             <div>
@@ -175,8 +171,8 @@ export default class Home extends React.Component {
             </div>
           </div>
 
-          <div className="mt60 flex-parent-mm">
-            <div className="flex-child flex-child--no-shrink mr24 mb0-mm mb18 w60 h60">
+          <div className="mt60 flex-mm">
+            <div className="flex-child-no-shrink mr24 mb0-mm mb18 w60 h60">
               <img src={defaultsImg} />
             </div>
             <div>
@@ -204,8 +200,8 @@ export default class Home extends React.Component {
             </div>
           </div>
 
-          <div className="mt60 flex-parent-mm">
-            <div className="flex-child flex-child--no-shrink mr24 mb0-mm mb18 w60 h60">
+          <div className="mt60 flex-mm">
+            <div className="flex-child-no-shrink mr24 mb0-mm mb18 w60 h60">
               <img src={customIconsImg} />
             </div>
             <div>
@@ -220,8 +216,8 @@ export default class Home extends React.Component {
             </div>
           </div>
 
-          <div className="mt60 flex-parent-mm">
-            <div className="flex-child flex-child--no-shrink mr24 mb0-mm mb18 w60 h60">
+          <div className="mt60 flex-mm">
+            <div className="flex-child-no-shrink mr24 mb0-mm mb18 w60 h60">
               <img src={mediaQueriesImg} />
             </div>
             <div>
@@ -264,8 +260,8 @@ export default class Home extends React.Component {
             </div>
           </div>
 
-          <div className="mt60 flex-parent-mm">
-            <div className="flex-child flex-child--no-shrink mr24 mb0-mm mb18 w60 h60">
+          <div className="mt60 flex-mm">
+            <div className="flex-child-no-shrink mr24 mb0-mm mb18 w60 h60">
               <img src={specificityImg} />
             </div>
             <div>
@@ -296,8 +292,8 @@ export default class Home extends React.Component {
             </div>
           </div>
 
-          <div className="mt60 flex-parent-mm">
-            <div className="flex-child flex-child--no-shrink mr24 mb0-mm mb18 w60 h60">
+          <div className="mt60 flex-mm">
+            <div className="flex-child-no-shrink mr24 mb0-mm mb18 w60 h60">
               <img src={boxModelImg} />
             </div>
             <div>
@@ -313,8 +309,8 @@ export default class Home extends React.Component {
             </div>
           </div>
 
-          <div className="mt60 flex-parent-mm">
-            <div className="flex-child flex-child--no-shrink mr24 mb0-mm mb18 w60 h60">
+          <div className="mt60 flex-mm">
+            <div className="flex-child-no-shrink mr24 mb0-mm mb18 w60 h60">
               <img src={isActiveImg} />
             </div>
             <div>
@@ -338,8 +334,8 @@ export default class Home extends React.Component {
             </div>
           </div>
 
-          <div className="mt60 flex-parent-mm">
-            <div className="flex-child flex-child--no-shrink mr24 mb0-mm mb18 w60 h60">
+          <div className="mt60 flex-mm">
+            <div className="flex-child-no-shrink mr24 mb0-mm mb18 w60 h60">
               <img src={focusImg} />
             </div>
             <div>